move trace_msg from tools.build_log to tools.utilities to avoid having to check value for 'trace' build option all the time
